Refactor and simplify `src/compiler`
