+++ src/compiler/CMakeLists.txt
@@ -1,22 +1,22 @@
 noa_library(NAMESPACE sourcemeta PROJECT jsonbinpack NAME compiler
   FOLDER "JSON BinPack/Compiler"
   SOURCES
-    states.h compiler.cc
-    mapper_rules/enum_8_bit.h
-    mapper_rules/enum_8_bit_top_level.h
-    mapper_rules/enum_arbitrary.h
-    mapper_rules/enum_singleton.h
-    mapper_rules/integer_bounded_8_bit.h
-    mapper_rules/integer_bounded_greater_than_8_bit.h
-    mapper_rules/integer_bounded_multiplier_8_bit.h
-    mapper_rules/integer_bounded_multiplier_greater_than_8_bit.h
-    mapper_rules/integer_lower_bound.h
-    mapper_rules/integer_lower_bound_multiplier.h
-    mapper_rules/integer_unbound.h
-    mapper_rules/integer_unbound_multiplier.h
-    mapper_rules/integer_upper_bound.h
-    mapper_rules/integer_upper_bound_multiplier.h
-    mapper_rules/number_arbitrary.h)
+    states.h encoding.h compiler.cc
+    mapper/enum_8_bit.h
+    mapper/enum_8_bit_top_level.h
+    mapper/enum_arbitrary.h
+    mapper/enum_singleton.h
+    mapper/integer_bounded_8_bit.h
+    mapper/integer_bounded_greater_than_8_bit.h
+    mapper/integer_bounded_multiplier_8_bit.h
+    mapper/integer_bounded_multiplier_greater_than_8_bit.h
+    mapper/integer_lower_bound.h
+    mapper/integer_lower_bound_multiplier.h
+    mapper/integer_unbound.h
+    mapper/integer_unbound_multiplier.h
+    mapper/integer_upper_bound.h
+    mapper/integer_upper_bound_multiplier.h
+    mapper/number_arbitrary.h)
 
 if(JSONBINPACK_INSTALL)
   noa_library_install(NAMESPACE sourcemeta PROJECT jsonbinpack NAME compiler)
