+++ src/compiler/mapper/enum_8_bit_top_level.h
@@ -1,14 +1,13 @@
 class Enum8BitTopLevel final : public sourcemeta::alterschema::Rule {
 public:
   Enum8BitTopLevel()
-      : sourcemeta::alterschema::Rule("enum_8_bit_top_level", "TODO") {};
+      : sourcemeta::alterschema::Rule{"enum_8_bit_top_level", ""} {};
 
   [[nodiscard]] auto condition(
       const sourcemeta::jsontoolkit::JSON &schema, const std::string &dialect,
       const std::set<std::string> &vocabularies,
       const sourcemeta::jsontoolkit::Pointer &pointer) const -> bool override {
-    return !is_encoding(schema) &&
-           dialect == "https://json-schema.org/draft/2020-12/schema" &&
+    return dialect == "https://json-schema.org/draft/2020-12/schema" &&
            vocabularies.contains(
                "https://json-schema.org/draft/2020-12/vocab/validation") &&
            schema.defines("enum") && schema.at("enum").is_array() &&
