+++ src/compiler/mapper/integer_lower_bound.h
@@ -1,29 +1,28 @@
-class IntegerUpperBound final : public sourcemeta::alterschema::Rule {
+class IntegerLowerBound final : public sourcemeta::alterschema::Rule {
 public:
-  IntegerUpperBound()
-      : sourcemeta::alterschema::Rule("integer_upper_bound", "TODO") {};
+  IntegerLowerBound()
+      : sourcemeta::alterschema::Rule{"integer_lower_bound", ""} {};
 
   [[nodiscard]] auto condition(const sourcemeta::jsontoolkit::JSON &schema,
                                const std::string &dialect,
                                const std::set<std::string> &vocabularies,
                                const sourcemeta::jsontoolkit::Pointer &) const
       -> bool override {
-    return !is_encoding(schema) &&
-           dialect == "https://json-schema.org/draft/2020-12/schema" &&
+    return dialect == "https://json-schema.org/draft/2020-12/schema" &&
            vocabularies.contains(
                "https://json-schema.org/draft/2020-12/vocab/validation") &&
            schema.defines("type") &&
            schema.at("type").to_string() == "integer" &&
-           !schema.defines("minimum") && schema.defines("maximum") &&
+           schema.defines("minimum") && !schema.defines("maximum") &&
            !schema.defines("multipleOf");
   }
 
   auto transform(sourcemeta::alterschema::Transformer &transformer) const
       -> void override {
-    auto maximum = transformer.schema().at("maximum");
+    auto minimum = transformer.schema().at("minimum");
     auto options = sourcemeta::jsontoolkit::JSON::make_object();
-    options.assign("maximum", std::move(maximum));
+    options.assign("minimum", std::move(minimum));
     options.assign("multiplier", sourcemeta::jsontoolkit::JSON{1});
-    make_encoding(transformer, "ROOF_MULTIPLE_MIRROR_ENUM_VARINT", options);
+    make_encoding(transformer, "FLOOR_MULTIPLE_ENUM_VARINT", options);
   }
 };
