+++ src/compiler/mapper/integer_unbound.h
@@ -1,24 +1,25 @@
-class EnumSingleton final : public sourcemeta::alterschema::Rule {
+class IntegerUnbound final : public sourcemeta::alterschema::Rule {
 public:
-  EnumSingleton() : sourcemeta::alterschema::Rule("enum_singleton", "TODO") {};
+  IntegerUnbound() : sourcemeta::alterschema::Rule{"integer_unbound", ""} {};
 
   [[nodiscard]] auto condition(const sourcemeta::jsontoolkit::JSON &schema,
                                const std::string &dialect,
                                const std::set<std::string> &vocabularies,
                                const sourcemeta::jsontoolkit::Pointer &) const
       -> bool override {
-    return !is_encoding(schema) &&
-           dialect == "https://json-schema.org/draft/2020-12/schema" &&
+    return dialect == "https://json-schema.org/draft/2020-12/schema" &&
            vocabularies.contains(
                "https://json-schema.org/draft/2020-12/vocab/validation") &&
-           schema.defines("enum") && schema.at("enum").is_array() &&
-           schema.at("enum").size() == 1;
+           schema.defines("type") &&
+           schema.at("type").to_string() == "integer" &&
+           !schema.defines("minimum") && !schema.defines("maximum") &&
+           !schema.defines("multipleOf");
   }
 
   auto transform(sourcemeta::alterschema::Transformer &transformer) const
       -> void override {
     auto options = sourcemeta::jsontoolkit::JSON::make_object();
-    options.assign("value", transformer.schema().at("enum").at(0));
-    make_encoding(transformer, "CONST_NONE", options);
+    options.assign("multiplier", sourcemeta::jsontoolkit::JSON{1});
+    make_encoding(transformer, "ARBITRARY_MULTIPLE_ZIGZAG_VARINT", options);
   }
 };
