+++ src/compiler/mapper/integer_unbound_multiplier.h
@@ -1,15 +1,14 @@
 class IntegerUnboundMultiplier final : public sourcemeta::alterschema::Rule {
 public:
   IntegerUnboundMultiplier()
-      : sourcemeta::alterschema::Rule("integer_unbound_multiplier", "TODO") {};
+      : sourcemeta::alterschema::Rule{"integer_unbound_multiplier", ""} {};
 
   [[nodiscard]] auto condition(const sourcemeta::jsontoolkit::JSON &schema,
                                const std::string &dialect,
                                const std::set<std::string> &vocabularies,
                                const sourcemeta::jsontoolkit::Pointer &) const
       -> bool override {
-    return !is_encoding(schema) &&
-           dialect == "https://json-schema.org/draft/2020-12/schema" &&
+    return dialect == "https://json-schema.org/draft/2020-12/schema" &&
            vocabularies.contains(
                "https://json-schema.org/draft/2020-12/vocab/validation") &&
            schema.defines("type") &&
